+++ universal-application-tool-0.0.1/app/controllers/applicant/ApplicantProgramBlocksController.java
@@ -28,7 +28,7 @@ import repository.ApplicationRepository;
 import services.applicant.ApplicantService;
 import services.applicant.Block;
 import services.applicant.ReadOnlyApplicantProgramService;
-import services.program.ProgramBlockNotFoundException;
+import services.applicant.ProgramBlockNotFoundException;
 import services.program.ProgramNotFoundException;
 import views.applicant.ApplicantProgramBlockEditView;
